requirements.txt
pillow
torch
matplotlib
random